requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
amqp
async-timeout
celery
click
click-didyoumean
click-plugins
click-repl
et_xmlfile
kombu
numpy
openpyxl
pandas
prompt_toolkit
psycopg2
python-dateutil
redis
six
tzdata
vine
wcwidth
